Adding Copy-Paste augmentation to mask-rcnn.
PiperOrigin-RevId: 367473178

+++ models/official/detection/utils/input_utils.py
@@ -509,8 +509,8 @@ def resize_and_crop_masks(masks,
   with tf.name_scope('resize_and_crop_masks'):
     mask_size = tf.cast(tf.shape(masks)[1:3], tf.float32)
     # Pad masks to avoid empty mask annotations.
-    masks = tf.concat([tf.zeros([1, mask_size[0], mask_size[1], 1]),
-                       masks], axis=0)
+    masks = tf.concat([tf.zeros([1, mask_size[0], mask_size[1], 1],
+                                dtype=masks.dtype), masks], axis=0)
 
     scaled_size = tf.cast(image_scale * mask_size, tf.int32)
     scaled_masks = tf.image.resize_images(
